split data loading functions from Swarm.Util
