+++ src/Swarm/DocGen.hs
@@ -56,7 +56,8 @@ import Swarm.Language.Syntax (Const (..))
 import Swarm.Language.Syntax qualified as Syntax
 import Swarm.Language.Typecheck (inferConst)
 import Swarm.TUI.Model.FailureRender qualified as F
-import Swarm.Util (getDataFileNameSafe, isRightOr, listEnums, quote)
+import Swarm.Util (isRightOr, listEnums, quote)
+import Swarm.Util.GameData (getDataFileNameSafe)
 import Text.Dot (Dot, NodeId, (.->.))
 import Text.Dot qualified as Dot
 import Witch (from)
