+++ src/Swarm/Game/Recipe.hs
@@ -51,7 +51,7 @@ import Data.Yaml
 import GHC.Generics (Generic)
 import Swarm.Game.Entity as E
 import Swarm.TUI.Model.Failure
-import Swarm.Util
+import Swarm.Util.GameData (getDataFileNameSafe)
 import Swarm.Util.Yaml
 import Witch
