+++ src/Swarm/Game/Scenario.hs
@@ -70,7 +70,7 @@ import Swarm.Game.Scenario.WorldDescription
 import Swarm.Language.Pipeline (ProcessedTerm)
 import Swarm.TUI.Model.Failure
 import Swarm.TUI.Model.FailureRender
-import Swarm.Util (getDataFileNameSafe)
+import Swarm.Util.GameData (getDataFileNameSafe)
 import Swarm.Util.Yaml
 import System.Directory (doesFileExist)
 import System.FilePath ((<.>), (</>))
