+++ src/Swarm/Game/State.hs
@@ -177,7 +177,8 @@ import Swarm.Language.Types
 import Swarm.Language.Value (Value)
 import Swarm.TUI.Model.Failure
 import Swarm.TUI.Model.FailureRender
-import Swarm.Util (getDataFileNameSafe, uniq, (<+=), (<<.=), (?))
+import Swarm.Util (uniq, (<+=), (<<.=), (?))
+import Swarm.Util.GameData (getDataFileNameSafe)
 import System.Clock qualified as Clock
 import System.Random (StdGen, mkStdGen, randomRIO)
