+++ src/Swarm/TUI/Controller.hs
@@ -101,6 +101,7 @@ import Swarm.TUI.Model.UI
 import Swarm.TUI.View (generateModal)
 import Swarm.TUI.View.Objective qualified as GR
 import Swarm.Util hiding ((<<.=))
+import Swarm.Util.GameData (getSwarmHistoryPath)
 import Swarm.Version (NewReleaseFailure (..))
 import System.Clock
 import System.FilePath (splitDirectories)
