+++ src/Swarm/TUI/Model/UI.hs
@@ -74,6 +74,7 @@ import Swarm.TUI.Model.Menu
 import Swarm.TUI.Model.Name
 import Swarm.TUI.Model.Repl
 import Swarm.Util
+import Swarm.Util.GameData (getSwarmHistoryPath, readAppData)
 import System.Clock
 
 ------------------------------------------------------------
