+++ src/Swarm/Util.hs
@@ -24,14 +24,6 @@ module Swarm.Util (
   -- * Directory utilities
   readFileMay,
   readFileMayT,
-  getSwarmXdgDataSubdir,
-  getSwarmXdgDataFile,
-  getSwarmSavePath,
-  getSwarmHistoryPath,
-  readAppData,
-  getDataDirSafe,
-  getDataFileNameSafe,
-  dataNotFound,
 
   -- * Text utilities
   isIdentChar,
@@ -71,14 +63,10 @@ module Swarm.Util (
 ) where
 
 import Control.Algebra (Has)
-import Control.Arrow (left)
 import Control.Effect.State (State, modify, state)
 import Control.Effect.Throw (Throw, throwError)
-import Control.Exception (catch)
-import Control.Exception.Base (IOException)
 import Control.Lens (ASetter', Lens', LensLike, LensLike', Over, lens, (<>~))
-import Control.Monad (forM, unless, when)
-import Control.Monad.Except (ExceptT (..), MonadIO, liftIO)
+import Control.Monad (unless)
 import Data.Bifunctor (first)
 import Data.Char (isAlphaNum)
 import Data.Either.Validation
@@ -87,7 +75,7 @@ import Data.List.NonEmpty (NonEmpty (..))
 import Data.List.NonEmpty qualified as NE
 import Data.Map (Map)
 import Data.Map qualified as M
-import Data.Maybe (fromMaybe, mapMaybe)
+import Data.Maybe (fromMaybe)
 import Data.Ord (comparing)
 import Data.Set (Set)
 import Data.Set qualified as S
@@ -100,20 +88,8 @@ import Language.Haskell.TH
 import Language.Haskell.TH.Syntax (lift)
 import NLP.Minimorph.English qualified as MM
 import NLP.Minimorph.Util ((<+>))
-import Paths_swarm (getDataDir)
-import Swarm.TUI.Model.Failure
 import System.Clock (TimeSpec)
-import System.Directory (
-  XdgDirectory (XdgData),
-  createDirectoryIfMissing,
-  doesDirectoryExist,
-  doesFileExist,
-  getXdgDirectory,
-  listDirectory,
- )
-import System.FilePath
 import System.IO.Error (catchIOError)
-import Witch
 
 infixr 1 ?
 infix 4 %%=, <+=, <%=, <<.=, <>=
@@ -187,97 +163,6 @@ readFileMayT = catchIO . T.readFile
 -- | Turns any IO error into Nothing.
 catchIO :: IO a -> IO (Maybe a)
 catchIO act = (Just <$> act) `catchIOError` (\_ -> return Nothing)
-
--- | Get subdirectory from swarm data directory.
---
--- This will first look in Cabal generated path and then
--- try a `data` directory in 'XdgData' path.
---
--- The idea is that when installing with Cabal/Stack the first
--- is preferred, but when the players install a binary they
--- need to extract the `data` archive to the XDG directory.
-getDataDirSafe :: MonadIO m => FilePath -> m (Either PathLoadFailure FilePath)
-getDataDirSafe p = do
-  d <- (`appDir` p) <$> liftIO getDataDir
-  de <- liftIO $ doesDirectoryExist d
-  if de
-    then return $ Right d
-    else do
-      xd <- (`appDir` p) <$> liftIO (getSwarmXdgDataSubdir False "data")
-      xde <- liftIO $ doesDirectoryExist xd
-      return $ if xde then Right xd else Left $ PathLoadFailure xd $ DoesNotExist Directory
- where
-  appDir r = \case
-    "" -> r
-    "." -> r
-    d -> r </> d
-
--- | Get file from swarm data directory.
---
--- See the note in 'getDataDirSafe'.
-getDataFileNameSafe ::
-  MonadIO m =>
-  FilePath ->
-  ExceptT PathLoadFailure m FilePath
-getDataFileNameSafe name = do
-  d <- ExceptT $ getDataDirSafe "."
-  let fp = d </> name
-  fe <- liftIO $ doesFileExist fp
-  ExceptT $
-    return $
-      if fe
-        then Right fp
-        else Left $ PathLoadFailure fp $ DoesNotExist File
-
--- | Get a nice message suggesting to download `data` directory to 'XdgData'.
-dataNotFound :: FilePath -> IO LoadingFailure
-dataNotFound f = do
-  d <- getSwarmXdgDataSubdir False ""
-  let squotes = squote . T.pack
-  return $
-    CustomMessage $
-      T.unlines
-        [ "Could not find the data: " <> squotes f
-        , "Try downloading the Swarm 'data' directory to: " <> squotes (d </> "data")
-        ]
-
--- | Get path to swarm data, optionally creating necessary
---   directories. This could fail if user has bad permissions
---   on his own $HOME or $XDG_DATA_HOME which is unlikely.
-getSwarmXdgDataSubdir :: Bool -> FilePath -> IO FilePath
-getSwarmXdgDataSubdir createDirs subDir = do
-  swarmData <- (</> subDir) <$> getXdgDirectory XdgData "swarm"
-  when createDirs (createDirectoryIfMissing True swarmData)
-  pure swarmData
-
-getSwarmXdgDataFile :: Bool -> FilePath -> IO FilePath
-getSwarmXdgDataFile createDirs filepath = do
-  let (subDir, file) = splitFileName filepath
-  d <- getSwarmXdgDataSubdir createDirs subDir
-  return $ d </> file
-
--- | Get path to swarm saves, optionally creating necessary
---   directories.
-getSwarmSavePath :: Bool -> IO FilePath
-getSwarmSavePath createDirs = getSwarmXdgDataSubdir createDirs "saves"
-
--- | Get path to swarm history, optionally creating necessary
---   directories.
-getSwarmHistoryPath :: Bool -> IO FilePath
-getSwarmHistoryPath createDirs = getSwarmXdgDataFile createDirs "history"
-
--- | Read all the .txt files in the data/ directory.
-readAppData :: ExceptT SystemFailure IO (Map Text Text)
-readAppData = do
-  d <- ExceptT $ left (AssetNotLoaded $ Data AppAsset) <$> getDataDirSafe "."
-  dirMembers <-
-    ExceptT $
-      fmap pure (listDirectory d) `catch` \(e :: IOException) ->
-        return $ Left $ AssetNotLoaded (Data AppAsset) $ PathLoadFailure d $ CustomMessage $ T.pack $ show e
-  let fs = filter ((== ".txt") . takeExtension) dirMembers
-
-  filesList <- liftIO $ forM fs (\f -> (into @Text (dropExtension f),) <$> readFileMayT (d </> f))
-  return $ M.fromList . mapMaybe sequenceA $ filesList
 
 ------------------------------------------------------------
 -- Some Text-y stuff
